rename log option
